pyproject.toml
[project]
name = "danish-economy"
version = "0.1.0"
description = "Danish Public Economy Explorer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "duckdb>=1.1",
    "polars>=1.0",
    "httpx>=0.28",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "black>=24.0",
    "mypy>=1.13",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/danish_economy"]

[tool.ruff]
src = ["src"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 88

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["danish_economy"]
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]
